--- client/components/NavBar/BottomNavBar/BottomNavBar.scss
@@ -1,35 +1,0 @@
-.bottom-navbar {
-  height: 100px;
-  width: 100%;
-  position: fixed !important;
-  bottom: 0;
-  background: #f2f2f2;
-
-  .navbar-nav {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-  }
-
-  .nav-item {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 25%;
-
-    .navbar-button {
-      width: 20px;
-      height: 20px;
-      padding: 0;
-      border: none;
-      background-color: #e0e0e0;
-    }
-
-    p {
-      font-size: 12px;
-      line-height: 12px;
-    }
-  }
-}
