--- client/components/NavBar/BottomNavBar/index.jsx
@@ -1,3 +1,0 @@
-import BottomNavBar from "./BottomNavBar";
-
-export default BottomNavBar;
