--- client/components/NavBar/TopNavBar/TopNavBar.scss
@@ -1,30 +1,0 @@
-.top-navbar {
-  height: 48px;
-  width: 100%;
-  position: fixed !important;
-  top: 0;
-  padding: 0 !important;
-  z-index: 999;
-
-  .navbar-nav {
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    background-color: #f2f2f2;
-    padding: 0 16px;
-  }
-
-  .nav-item {
-    .navbar-logo {
-      width: 124px;
-      align-self: center !important;
-    }
-
-    .navbar-avatar {
-      width: 24px;
-      border-radius: 50%;
-      align-self: flex-start;
-    }
-  }
-}
