--- client/components/NavBar/TopNavBar/index.jsx
@@ -1,3 +1,0 @@
-import TopNavBar from "./TopNavBar";
-
-export default TopNavBar;
