+++ client/components/NavBar/navbar.module.scss
@@ -1,0 +1,42 @@
+@import "../../../public/static/colors.scss";
+
+// Colors
+$btn: #e0e0e0;
+$active-btn: #bdbdbd;
+
+// Top Nav
+.Logo {
+  width: 124px;
+}
+
+.Avatar {
+  width: 24px;
+  border-radius: 50%;
+}
+
+// Bottom Nav
+.BottomNav {
+  position: fixed;
+  bottom: 0;
+  height: 100px;
+  width: 100%;
+  background: $background;
+  display: flex;
+  justify-content: space-between;
+
+  .NavItem {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 25%;
+
+    .NavButton {
+      width: 20px;
+      height: 20px;
+      border-radius: 5px;
+      background-color: $btn;
+      margin-bottom: 6px;
+    }
+  }
+}
