+++ public/static/App.scss
@@ -72,6 +72,6 @@ h6 {
 }
 
 p {
-  font-size: 22px;
-  line-height: 30px;
+  font-size: 12px;
+  line-height: 12px;
 }
